chore(types): split out lint and type commands to be per-package (#9032)
* stash

* stash progress on lint and tsc split out

* cleanup

* more cleanup

* attempt to fix deps

* get more things working

* more lockfile cleanup

* things

* fix builds

* fix test

* port turbo work

* small tweaks

* fix boolean

* decycle

* fix another cycle

* build

* fix build

* more improvements

* fix main tests

* more fixups

* remove some dead stuff

* remove unused pnpm peer ignores

* cleanup test commands

* more task cleanup

* fixup some tests

* some turbo cleanup

* more command cleanup

* make tests nicer?

* fix fastboot tests

* gets it all working

* test:production

* fix test:production

* fixup some test runs

* skip adapter prod

* more test fixes for prod

+++ packages/diagnostic/server/bun/socket-handler.js
@@ -1,6 +1,7 @@
-import { info, debug } from "../utils/debug.js";
 import chalk from 'chalk';
-import { sinceStart } from "../utils/time.js";
+
+import { debug, info } from '../utils/debug.js';
+import { sinceStart } from '../utils/time.js';
 
 export function buildHandler(config, state) {
   return {
@@ -27,24 +28,33 @@ export function buildHandler(config, state) {
         case 'suite-finish':
           config.reporter.onSuiteFinish(msg);
 
-          ws.send(JSON.stringify({ name: 'close' }));
-          ws.close();
+          if (!config.serve) {
+            ws.send(JSON.stringify({ name: 'close' }));
+            ws.close();
+          }
           state.completed++;
-          debug(`${chalk.green('✅ [Complete]')} ${chalk.cyan(msg.browserId)}/${chalk.cyan(msg.windowId)} ${chalk.yellow('@' + sinceStart())}`);
+          debug(
+            `${chalk.green('✅ [Complete]')} ${chalk.cyan(msg.browserId)}/${chalk.cyan(msg.windowId)} ${chalk.yellow(
+              '@' + sinceStart()
+            )}`
+          );
           if (state.completed === state.expected) {
             const exitCode = config.reporter.onRunFinish(msg);
             debug(`${chalk.green('✅ [All Complete]')} ${chalk.yellow('@' + sinceStart())}`);
-            state.browsers.forEach((browser) => {
-              browser.proc.kill();
-              browser.proc.unref();
-            });
-            state.server.stop();
-            if (config.cleanup) {
-              debug(`Running configured cleanup hook`);
-              await config.cleanup();
-              debug(`Configured cleanup hook completed`);
+
+            if (!config.serve) {
+              state.browsers.forEach((browser) => {
+                browser.proc.kill();
+                browser.proc.unref();
+              });
+              state.server.stop();
+              if (config.cleanup) {
+                debug(`Running configured cleanup hook`);
+                await config.cleanup();
+                debug(`Configured cleanup hook completed`);
+              }
+              process.exit(exitCode);
             }
-            process.exit(exitCode);
           }
 
           break;
@@ -54,5 +64,5 @@ export function buildHandler(config, state) {
     open(ws) {}, // a socket is opened
     close(ws, code, message) {}, // a socket is closed
     drain(ws) {}, // the socket is ready to receive more data
-  }
-};
+  };
+}
